Theme provider updated, Some unnecessary functions removed also some conflicting portion updated
